+++ crates/egui-winit/src/lib.rs
@@ -594,6 +594,7 @@ impl State {
                 self.egui_input.events.push(egui::Event::Key {
                     key,
                     pressed,
+                    repeat: false, // egui will fill this in for us!
                     modifiers: self.egui_input.modifiers,
                 });
             }
@@ -654,8 +655,8 @@ impl State {
     }
 
     fn set_cursor_icon(&mut self, window: &winit::window::Window, cursor_icon: egui::CursorIcon) {
-        // prevent flickering near frame boundary when Windows OS tries to control cursor icon for window resizing
-        #[cfg(windows)]
+        // Prevent flickering near frame boundary when Windows OS tries to control cursor icon for window resizing.
+        // On other platforms: just early-out to save CPU.
         if self.current_cursor_icon == cursor_icon {
             return;
         }